pyproject.toml
[project]
name = "ticket-evaluator"
version = "1.0.0"
description = "LLM-based ticket reply evaluation system"
requires-python = ">=3.10"
dependencies = [
    "groq>=0.4.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
evaluate-tickets = "evaluate_tickets:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["evaluate_tickets.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
